+++ src/interfaces/IVault.sol
@@ -16,7 +16,7 @@ interface IVault {
     function initialize(
         address _owner,
         address registryAddress,
-        uint8 numeraire,
+        uint256 numeraire,
         address stable,
         address stakeContract,
         address interestModule
